+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0
 - Improve error message when a NuRoute is registered twice
+- Improve error handling on empty NuRouters
 - Add option to get registered routes after NuRouter initialization when overriding the `lazyRouteRegister` option in NuRouter
 - [BREAKING] Change the `NuRouter.onError` to handle any thrown object (including `Exception`)
   - Type of the `NuRouter.onError` function changed from `Exception` to `Object`